😲 7th Commit
Implemented ExpensesList component with dummy data.
Added a FlatList to display the expenses.
Each expense item is rendered using a separate function, renderExpenseItem, which currently just displays the description of the expense.
Also added keyExtractor prop to FlatList to use the unique ID of each expense as the key.
Note: The list of expenses is the same on both AllExpenses and RecentExpenses screens as filtering logic has not been implemented yet.