🤑 18th Commit
Implemented expensesReducer hook in expenses-context.js for managing expenses.
Added ADD, UPDATE, and DELETE cases in the reducer.
For ADD, a new expense object is created and added to the existing state array.
For UPDATE, the index of the expense to be updated is found, and a new object is created with the updated values, which replaces the old object in the state array.
For DELETE, the state array is filtered to exclude the expense to be deleted.
Ensured all state updates are done immutably.
Initialized the state with DUMMY_EXPENSES cut from ExpensesOutput.js file.