33rd Commit
Integrated axios for HTTP requests:
Created http.js utility file with storeExpense function for sending POST requests to Firebase.
Added axios to send POST requests to Firebase in storeExpense function.
Integrated storeExpense in confirmHandler function in ManageExpense.js, enabling expenses to be stored on Firebase when not in editing mode.