36th Commit
Added update and delete functionality for expenses: Created updateExpense and deleteExpense functions in http.js to send PUT and DELETE requests to Firebase respectively.
Integrated these functions in ManageExpense.js.
The updateExpense function is called with the expense ID and updated data when the user edits an expense.
The deleteExpense function is called with the expense ID when the user deletes an expense.
Both functions are awaited to ensure the request completes before the UI updates.